This commit adds information about how to use the ossearch module with OpenSim.
Added a README file and removed the OpenSimSearch/OpenSim-addition.ini.example
file as the information it contained is now in the new README file.

webroot/query.php:
The previous commit should have stated that results of a search of classifieds
are now returned in descending order based on the listing price.


git-svn-id: http://forge.opensimulator.org/svn/ossearch/trunk@161 109abde3-1f22-4c10-b5d7-b6b0135cd00c